archive/requirements.txt
# requirements.txt
Jinja2>=3.0
Markdown>=3.3
python-frontmatter>=1.0
PyYAML>=5.4 # Often a dependency for python-frontmatter